+++ packages/adapter-api/index.ts
@@ -25,3 +25,5 @@ export * from './src/change'
 export * from './src/change_group'
 export * from './src/utils'
 export * from './src/error'
+export * from './src/comparison'
+export * from './src/reference_comparison'
