+++ packages/adapter-api/src/values.ts
@@ -236,99 +236,6 @@ export const isTypeReference = (value: any): value is TypeReference => (
   value instanceof TypeReference
 )
 
-/*
-  Benchmarking reveals that looping on strings is extremely expensive.
-  It seems that random access to a string is, for some reason, a bit expensive.
-  Using "replace" takes about 30 times as much as a straightforward comparison.
-  However, it's about 20 times better to use replace than to iterate over both strings.
-  For this reason we first check a naive comparison, and then test with replace.
-  */
-const compareStringsIgnoreNewlineDifferences = (s1: string, s2: string): boolean =>
-  (s1 === s2) || (s1.replace(/\r\n/g, '\n') === s2.replace(/\r\n/g, '\n'))
-
-export const shouldResolve = (value: unknown): boolean => (
-  // We don't resolve references to elements because the logic of how to resolve each
-  // reference currently exists only in the adapter so here we don't know how to
-  // resolve them.
-  // We do resolve variables because they always point to a primitive value that we can compare.
-  // If a value is not a reference we decided to return that we should "resolve" it so
-  // the value will be treated like a resolved reference
-  !isReferenceExpression(value) || !value.elemID.isBaseID() || value.elemID.idType === 'var'
-)
-
-const shouldCompareByValue = (
-  first: Value,
-  second: Value,
-  options?: CompareOptions,
-): boolean => Boolean(options?.compareByValue)
-  && shouldResolve(first)
-  && shouldResolve(second)
-
-const compareSpecialValuesWithCircularRefs = (
-  first: Value,
-  second: Value,
-  comparedReferences: Set<string>,
-  options?: CompareOptions,
-): boolean | undefined => {
-  if (isStaticFile(first) && isStaticFile(second)) {
-    return first.isEqual(second)
-      && (options?.compareByValue
-        ? true
-        : first.filepath === second.filepath)
-  }
-  if (isReferenceExpression(first) || isReferenceExpression(second)) {
-    if (shouldCompareByValue(first, second, options)) {
-      if (isReferenceExpression(first) && isReferenceExpression(second)) {
-        const key = `${first.elemID.getFullName()}-${second.elemID.getFullName()}`
-        if (comparedReferences.has(key)) {
-          // We assume two circular references are equal (sames as _.isEqual with two circular objects)
-          return true
-        }
-
-        comparedReferences.add(key)
-      }
-
-      const fValue = isReferenceExpression(first) ? first.value : first
-      const sValue = isReferenceExpression(second) ? second.value : second
-
-      return _.isEqualWith(
-        fValue,
-        sValue,
-        (va1, va2) => compareSpecialValuesWithCircularRefs(va1, va2, comparedReferences, options),
-      )
-    }
-
-    if (isReferenceExpression(first) && isReferenceExpression(second)) {
-      return first.elemID.isEqual(second.elemID)
-    }
-    // If one side is a reference and the other is not and we should not
-    // compare by values then the values are different
-    return false
-  }
-  if (typeof first === 'string' && typeof second === 'string') {
-    return compareStringsIgnoreNewlineDifferences(first, second)
-  }
-  return undefined
-}
-
-
-export const compareSpecialValues = (
-  first: Value,
-  second: Value,
-  options?: CompareOptions,
-): boolean | undefined =>
-  compareSpecialValuesWithCircularRefs(first, second, new Set(), options)
-
-export const isEqualValues = (
-  first: Value,
-  second: Value,
-  options?: CompareOptions,
-): boolean => _.isEqualWith(
-  first,
-  second,
-  (va1, va2) => compareSpecialValues(va1, va2, options),
-)
-
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const isVariableExpression = (value: any): value is VariableExpression => (
   value instanceof VariableExpression
