+++ packages/workspace/src/validator.ts
@@ -499,10 +499,11 @@ const validateValue = (
     if (!isElement(value.value)
       // We check for both validatedReferences and isParentOf because if we got to
       // the same reference twice but not through a cycle we do want to validate it
-      && !(validatedReferences.has(value.elemID.getFullName())
-        && (value.elemID.isParentOf(elemID) || value.elemID.isEqual(elemID)))) {
+      && !validatedReferences.has(value.elemID.getFullName())) {
       validatedReferences.add(value.elemID.getFullName())
-      return validateValue(elemID, value.value, type, validatedReferences)
+      const result = validateValue(elemID, value.value, type, validatedReferences)
+      validatedReferences.delete(value.elemID.getFullName())
+      return result
     }
     return []
   }
